Ugly hack to support pagination URLs with non-default {CATEGORY,TAG}_SAVE_AS settings in pelicanconf.py
Prior to this change, with pagination enabled and with *_SAVE_AS = some
wordpress-like URL, the pagination URLs generated for any given tag or
category were completely broken, e.g.

    http://www.vcheng.org/tag/debian/.html
    http://www.vcheng.org/tag/debian/2.html
    http://www.vcheng.org/tag/debian/3.html
    [...]

What I want instead is to generate URLs like the following:

    http://www.vcheng.org/tag/debian/index.html
    http://www.vcheng.org/tag/debian/index2.html
    http://www.vcheng.org/tag/debian/index3.html
    [...]

However, doing so unconditionally breaks pagination on the actual index page
(as opposed to the tag/category pages), e.g.

    http://www.vcheng.org/indexindex.html
    http://www.vcheng.org/indexindex2.html (and so on)

So revert to the original behaviour if page_name == 'index'.

TODO: Make templates/modules/pagination.html more modular/flexible because
otherwise it'll break if I ever decide to change the URL format my tag and
category pages are saved as (including back to the defaults, so this can't be
forwarded upstream as-is).